Refactor : Paper의 정보 재사용을 위해 컴포넌트로 분리
- PaperInfo 부분 post와 list 페이지 모두에서 사용하므로 공통 컴포넌트로 분리
- PaperInfo 컴포넌트는 롤링페이퍼 데이터의 name, recentMessages, messageCount, topReactions를 props로 받아서 사용하도록 함
