+++ src/components/Paper/PaperBox.style.js
@@ -4,12 +4,12 @@ const Container = styled.div`
   display: flex;
   margin-top: 50px;
   flex-direction: column;
+  row-gap: 16px;
 
   span {
     font-family: Pretendard-Bold;
     font-weight: 700;
     font-size: 24px;
-    margin-bottom: 16px;
   }
 `;
